+++ package/src/components/Attachment/hooks/useLoadingImage.tsx
@@ -1,8 +1,0 @@
-import { useState } from 'react';
-
-export const useLoadingImage = () => {
-  const [isLoadingImage, setLoadingImage] = useState(true);
-  const [isLoadingImageError, setLoadingImageError] = useState(false);
-
-  return { isLoadingImage, isLoadingImageError, setLoadingImage, setLoadingImageError };
-};