+++ package/src/store/apis/getAllChannelIds.ts
@@ -1,0 +1,7 @@
+import { selectChannels } from './queries/selectChannels';
+
+export const getAllChannelIds = () => {
+  const channels = selectChannels();
+
+  return channels.map((c) => c.cid);
+};