+++ package/src/store/mappers/mapDateTimeToStorable.ts
@@ -1,0 +1,5 @@
+export const mapDateTimeToStorable = (datetime?: string) => {
+  if (!datetime) return '';
+
+  return new Date(datetime).toISOString();
+};